Switch to Sarabun font and update font usage
Replaced Geist and Geist_Mono fonts with Sarabun in the layout and updated global CSS to remove old font variables. Updated schedules page to use the new font and refactored schedules list to use a helper for page size. Also updated the build script to run 'prisma generate' and changed the favicon.